Работа по заданию 4.7 Модуляция
