+++ index.html
@@ -230,7 +230,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script defer src="js/main.js"></script>
+  <script defer src="js/main.js" type="module"></script>
 
 </body>
 </html>
